+++ R/ECgenerator.R
@@ -9,8 +9,11 @@
 #' @return Error code word polynomial
 #' @export
 #' @importFrom utils tail
+#' @author Victor Teh
+#' @family legacy
 
 ECgenerator <- function(GenPoly, DataPoly, DCWordCount, ECWordCount) { #nolint
+  .Deprecated("qr_code")
   logTable <- create_log_table()
 
   targetDataPoly <- c(DataPoly, rep(0, ECWordCount))