+++ R/qrFillUpMatrix.R
@@ -8,9 +8,12 @@
 #'
 #' @return matrix filled up with the data bits
 #' @export
+#' @author Victor Teh
+#' @family legacy
 
 
 qrFillUpMatrix <- function(allBinary, data, version) {
+  .Deprecated("qr_code")
   counter <- 1
   direction <- 1
   byteCount <- length(allBinary) / 8