+++ R/versionPolyGen.R
@@ -7,8 +7,11 @@
 #'
 #' @return version polynomial.
 #'
+#' @author Victor Teh
+#' @family legacy
 
 versionPolyGen <- function(versionString, polyString) {
+  .Deprecated("qr_code")
 
   versionString <- as.integer(unlist(strsplit(versionString, split = "")))
   oriVersionString <- versionString